requirements.txt
attrdict==2.0.1
joblib==1.4.2
loguru==0.7.3
matplotlib==3.10.1
numpy==2.2.5
pandas==2.2.3
python-dotenv==1.1.0
scikit_learn==1.6.1
scipy==1.15.2
torch==2.6.0
torch_geometric==2.6.1
tqdm==4.67.1
typer==0.15.3
